refactor: renamed test packages for amqp communication
