+++ SmartTutorTests/Unit/SmartTutorAMQPTests.cs
@@ -3,7 +3,7 @@ using Xunit;
 
 namespace SmartTutorTests.Unit
 {
-    public class SmartTutorServiceTest
+    public class SmartTutorAMQPTests
     {
         [Fact]
         public void Consume_Issue_Report_Message()
